remove old client logic and cleanup testing.
