+++ client/src/bucket.rs
@@ -153,95 +153,160 @@ fn write_to_read(write: &WriteRequest) -> ReadResult {
     ReadResult::new(write.kind(), write.id(), write.data())
 }
 
-#[test]
-fn test_bucket() {
+#[cfg(test)]
+mod tests {
+    use super::*;
     use crate::provider::Provider;
+    use std::iter::empty;
 
-    let key = Key::<Provider>::random().expect(line_error!());
+    #[test]
+    fn test_bucket() {
+        use crate::provider::Provider;
 
-    let mut bucket = Bucket::<Provider>::new();
-    let id1 = RecordId::random::<Provider>().expect(line_error!());
-    let id2 = RecordId::random::<Provider>().expect(line_error!());
+        let key = Key::<Provider>::random().expect(line_error!());
 
-    bucket.take(key.clone(), |view, mut reads| -> Vec<ReadResult> {
-        let mut writer = view.writer(id1);
+        let mut bucket = Bucket::<Provider>::new();
+        let id1 = RecordId::random::<Provider>().expect(line_error!());
+        let id2 = RecordId::random::<Provider>().expect(line_error!());
 
-        let wr = writer.truncate().expect(line_error!());
+        bucket.take(key.clone(), |view, mut reads| -> Vec<ReadResult> {
+            let mut writer = view.writer(id1);
 
-        reads.push(write_to_read(&wr));
+            let wr = writer.truncate().expect(line_error!());
 
-        let wr = writer
-            .write(b"some data", RecordHint::new(b"").expect(line_error!()))
-            .expect(line_error!());
+            reads.push(write_to_read(&wr));
 
-        let mut wr: Vec<ReadResult> = wr.into_iter().map(|w| write_to_read(&w)).collect();
+            let wr = writer
+                .write(b"some data", RecordHint::new(b"").expect(line_error!()))
+                .expect(line_error!());
 
-        reads.append(&mut wr);
+            let mut wr: Vec<ReadResult> = wr.into_iter().map(|w| write_to_read(&w)).collect();
 
-        let mut writer = view.writer(id2);
+            reads.append(&mut wr);
 
-        let wr = writer.truncate().expect(line_error!());
+            let mut writer = view.writer(id2);
 
-        reads.push(write_to_read(&wr));
+            let wr = writer.truncate().expect(line_error!());
 
-        reads
-    });
+            reads.push(write_to_read(&wr));
 
-    bucket.take(key.clone(), |view, reads| {
-        let data: Vec<(RecordId, RecordHint)> = view.records().collect();
-
-        data.iter().for_each(|(i, h)| {
-            println!("{:?}: {:?}", i, h);
+            reads
         });
 
-        reads
-    });
+        bucket.take(key.clone(), |view, reads| {
+            let data: Vec<(RecordId, RecordHint)> = view.records().collect();
 
-    bucket.take(key.clone(), |view, reads| {
-        let reader = view.reader();
+            data.iter().for_each(|(i, h)| {
+                println!("{:?}: {:?}", i, h);
+            });
+
+            reads
+        });
+
+        bucket.take(key.clone(), |view, reads| {
+            let reader = view.reader();
+
+            let res = reader.prepare_read(&id1).expect(line_error!());
+
+            match res {
+                PreparedRead::CacheHit(v) => {
+                    println!("{:?}", std::str::from_utf8(&v));
+                }
+                PreparedRead::CacheMiss(v) => {
+                    println!("{:?}", v.id());
+                }
+                _ => {
+                    println!("no data");
+                }
+            }
+
+            reads
+        });
+
+        bucket.take(key.clone(), |view, mut reads| {
+            let mut writer = view.writer(id1);
+
+            reads.push(write_to_read(&writer.revoke().expect(line_error!())));
+
+            reads
+        });
+
+        bucket.take(key.clone(), |view, mut reads| {
+            let deletes = view.gc();
+
+            deletes.iter().for_each(|d| {
+                reads.retain(|r| r.id() != d.id());
+            });
+
+            reads
+        });
+
+        bucket.take(key, |view, reads| {
+            let data: Vec<(RecordId, RecordHint)> = view.records().collect();
+
+            data.iter().for_each(|(i, h)| {
+                println!("Data {:?}: {:?}", i, h);
+            });
+
+            reads
+        });
+    }
+
+    #[test]
+    fn test_stuff() {
+        let id1 = RecordId::random::<Provider>().expect(line_error!());
+        let id2 = RecordId::random::<Provider>().expect(line_error!());
+        let key1 = Key::<Provider>::random().expect(line_error!());
+        let key2 = Key::<Provider>::random().expect(line_error!());
+        let mut map: HashMap<Key<Provider>, Option<DBView<Provider>>> = HashMap::new();
+        let mut writes1 = vec![];
+        let mut writes2 = vec![];
+        let view1 = DBView::load(key1.clone(), empty::<ReadResult>()).expect(line_error!());
+        let view2 = DBView::load(key2.clone(), empty::<ReadResult>()).expect(line_error!());
+        let mut writer1 = view1.writer(id1);
+        let mut writer2 = view2.writer(id2);
+        writes1.push(writer1.truncate().expect(line_error!()));
+        writes2.push(writer2.truncate().expect(line_error!()));
+        writes1.append(
+            &mut writer1
+                .write(b"some data", RecordHint::new(b"").expect(line_error!()))
+                .expect(line_error!()),
+        );
+        writes2.append(
+            &mut writer2
+                .write(b"some data", RecordHint::new(b"").expect(line_error!()))
+                .expect(line_error!()),
+        );
+        let view1 = DBView::load(key1.clone(), writes1.iter().map(write_to_read)).expect(line_error!());
+        let view2 = DBView::load(key2.clone(), writes2.iter().map(write_to_read)).expect(line_error!());
+        map.insert(key1.clone(), Some(view1));
+        map.insert(key2.clone(), Some(view2));
+        let view1 = map.remove(&key1).unwrap().unwrap();
+        let id12 = RecordId::random::<Provider>().expect(line_error!());
+        let id22 = RecordId::random::<Provider>().expect(line_error!());
+        let mut writer1 = view1.writer(id12);
+        writes1.push(writer1.truncate().expect(line_error!()));
+        writes1.append(
+            &mut writer1
+                .write(b"some data", RecordHint::new(b"").expect(line_error!()))
+                .expect(line_error!()),
+        );
+        writes1.push(writer1.truncate().expect(line_error!()));
+        let view1 = DBView::load(key1.clone(), writes1.iter().map(write_to_read)).expect(line_error!());
+        map.insert(key1.clone(), Some(view1));
+
+        let view1 = DBView::load(key1.clone(), writes1.iter().map(write_to_read)).expect(line_error!());
+
+        let reader = view1.reader();
 
         let res = reader.prepare_read(&id1).expect(line_error!());
 
         match res {
-            PreparedRead::CacheHit(v) => {
-                println!("{:?}", std::str::from_utf8(&v));
-            }
-            PreparedRead::CacheMiss(v) => {
-                println!("{:?}", v.id());
-            }
-            _ => {
-                println!("no data");
-            }
+            PreparedRead::CacheHit(v) => println!("{:?}", std::str::from_utf8(&v)),
+            _ => println!("no data"),
         }
-
-        reads
-    });
-
-    bucket.take(key.clone(), |view, mut reads| {
-        let mut writer = view.writer(id1);
-
-        reads.push(write_to_read(&writer.revoke().expect(line_error!())));
-
-        reads
-    });
-
-    bucket.take(key.clone(), |view, mut reads| {
-        let deletes = view.gc();
-
-        deletes.iter().for_each(|d| {
-            reads.retain(|r| r.id() != d.id());
-        });
-
-        reads
-    });
-
-    bucket.take(key, |view, reads| {
-        let data: Vec<(RecordId, RecordHint)> = view.records().collect();
-
-        data.iter().for_each(|(i, h)| {
-            println!("Data {:?}: {:?}", i, h);
-        });
-
-        reads
-    });
+    }
+    fn write_to_read(write: &WriteRequest) -> ReadResult {
+        ReadResult::new(write.kind(), write.id(), write.data())
+    }
 }
